refactor: refactor validator classes and update usage
The 'validator.ts' has been renamed to 'siretValidator.ts' and the class name along with its method has been updated. The code also includes changes in importing the renamed validator in various files. Further, 'phoneValidator.ts' and 'emailValidator.ts' have been moved from web to common package.
