+++ packages/backend/src/establishment/domain/establishmentFeatures.ts
@@ -1,7 +1,7 @@
 import { Result } from 'true-myth'
 import type { CityToRegionMapping, EstablishmentRepository, NafMapping } from './spi'
 import { Establishment, EstablishmentDetails, EstablishmentSearch, SearchResult, EstablishmentFront, Siret } from './types'
-import Validator from '../../../../common/src/establishment/validator'
+import SiretValidator from '@tee/common/src/establishment/validator/siretValidator'
 
 export default class EstablishmentFeatures {
   private readonly _establishmentRepository: EstablishmentRepository
@@ -15,7 +15,7 @@ export default class EstablishmentFeatures {
   }
 
   public async search(query: string): Promise<Result<EstablishmentSearch, Error>> {
-    if (Validator.validateSiret(query)) {
+    if (SiretValidator.validate(query)) {
       const bySiretResult = await this._searchBySiret(query)
       if (bySiretResult.isOk) {
         return bySiretResult
