+++ packages/web/src/components/track/form/TrackSiret.vue
@@ -117,7 +117,7 @@ import Matomo from '@/utils/matomo'
 import Navigation from '@/utils/navigation'
 import TrackSiret from '@/utils/track/TrackSiret'
 import Translation from '@/utils/translation'
-import Validator from '@tee/common/src/establishment/validator'
+import SiretValidator from '@tee/common/src/establishment/validator/siretValidator'
 import { ref, computed } from 'vue'
 import { EstablishmentSearch } from '@/types'
 
@@ -173,7 +173,7 @@ const processInput = async () => {
 
   if (!queryValue.value || queryValue.value.length < 3) {
     errorMessage.value = Translation.t('enterprise.searchTooShort')
-  } else if (Validator.isValidSiretFormat(queryValue.value) && !Validator.isValidSiretNumber(queryValue.value)) {
+  } else if (SiretValidator.isValidSiretFormat(queryValue.value) && !SiretValidator.isValidSiretNumber(queryValue.value)) {
     errorMessage.value = "Le numéro SIRET n'est pas valide"
   } else {
     const searchResult = await TrackSiret.search(queryValue.value)
